VITIS-6253 Refactor update base into seperate option class
